+++ Assets/Puppet/PuppetTransmission.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 55c7e56bb80a0479d9c144b042144485
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
